Separating test 2

+++ aiida/backends/djsite/db/testbase.py
@@ -5,6 +5,7 @@ Base class for AiiDA tests
 from django.utils import unittest
 import shutil
 import os
+from aiida.backends.testimplbase import AiidaTestImplementation
 
 # Add a new entry here if you add a file with tests under aiida.backends.djsite.db.subtests
 # The key is the name to use in the 'verdi test' command (e.g., a key 'generic'
@@ -16,83 +17,50 @@ __license__ = "MIT license, see LICENSE.txt file."
 __version__ = "0.7.0"
 __authors__ = "The AiiDA team."
 
-db_test_list = {
-    'generic': ['aiida.backends.djsite.db.subtests.generic'],
-    'nodes': ['aiida.backends.djsite.db.subtests.nodes'],
-    'nwchem': ['aiida.backends.djsite.db.subtests.nwchem'],
-    'dataclasses': ['aiida.backends.djsite.db.subtests.dataclasses'],
-    'qepw': ['aiida.backends.djsite.db.subtests.quantumespressopw'],
-    'codtools': ['aiida.backends.djsite.db.subtests.codtools'],
-    'dbimporters': ['aiida.backends.djsite.db.subtests.dbimporters'],
-    'export_and_import': ['aiida.backends.djsite.db.subtests.export_and_import'],
-    'migrations': ['aiida.backends.djsite.db.subtests.migrations'],
-    'parsers': ['aiida.backends.djsite.db.subtests.parsers'],
-    'qepwinputparser': ['aiida.backends.djsite.db.subtests.pwinputparser'],
-    'qepwimmigrant': ['aiida.backends.djsite.db.subtests.quantumespressopwimmigrant'],
-    'tcodexporter': ['aiida.backends.djsite.db.subtests.tcodexporter'],
-    'workflows': ['aiida.backends.djsite.db.subtests.workflows'],
-    'query': ['aiida.backends.djsite.db.subtests.query'],
-    'backup': ['aiida.backends.djsite.db.subtests.backup_script',
-               'aiida.backends.djsite.db.subtests.backup_setup_script'],
-    'restapi' : ['aiida.backends.djsite.db.subtests.restapi'],
-    'calculation_node': ['aiida.backends.djsite.db.subtests.calculation_node'],
-}
-
-
-class AiidaTestCase(unittest.TestCase):
+# This contains the codebase for the setUpClass and tearDown methods used internally by the AiidaTestCase
+# This inherits only from 'object' to avoid that it is picked up by the automatic discovery of tests
+# (It shouldn't, as it risks to destroy the DB if there are not the checks in place, and these are
+# implemented in the AiidaTestCase
+class DjangoTests(AiidaTestImplementation):
     """
     Automatically takes care of the setUpClass and TearDownClass, when needed.
     """
 
-    @classmethod
-    def setUpClass(cls, initial_data=True):
+    # Note this is has to be a normal method, not a class method
+    def setUpClass_method(self):
+        self.clean_db()
+        self.insert_data()
 
+    def insert_data(self):
+        """
+        Insert default data in DB.
+        """
         from django.core.exceptions import ObjectDoesNotExist
 
         from aiida.backends.djsite.db.models import DbUser
         from aiida.orm.computer import Computer
         from aiida.common.utils import get_configured_user_email
-
-        # # Check if the database is initially empty
-        # from aiida.backends.djsite.db.models import DbComputer
-        # if DbComputer.objects.exists():
-        #     print "database not empty"
-        #     exit(0)
-
         # We create the user only once:
         # Otherwise, get_automatic_user() will fail when the
         # user is recreated because it caches the user!
-        # In any case, store it in cls.user though
-        # Other possibility: flush the user cache on delete
-        if initial_data:
-            try:
-                cls.user = DbUser.objects.get(email=get_configured_user_email())
-            except ObjectDoesNotExist:
-                cls.user = DbUser.objects.create_user(get_configured_user_email(),
-                                                      'fakepwd')
-            cls.computer = Computer(name='localhost',
-                                    hostname='localhost',
-                                    transport_type='local',
-                                    scheduler_type='pbspro',
-                                    workdir='/tmp/aiida')
-            cls.computer.store()
+        # In any case, store it in self.user though
+        try:
+            self.user = DbUser.objects.get(email=get_configured_user_email())
+        except ObjectDoesNotExist:
+            self.user = DbUser.objects.create_user(get_configured_user_email(),
+                                                  'fakepwd')
+        ## Reqired by the calling class
+        self.user_email = self.user.email
 
-    @classmethod
-    def tearDownClass(cls):
-        # exit(0)
-        from aiida.settings import REPOSITORY_PATH
-        from aiida.common.setup import TEST_KEYWORD
-        from aiida.common.exceptions import InvalidOperation
+        ## Also self.computer is required by the calling class
+        self.computer = Computer(name='localhost',
+                                hostname='localhost',
+                                transport_type='local',
+                                scheduler_type='pbspro',
+                                workdir='/tmp/aiida')
+        self.computer.store()
 
-        base_repo_path = os.path.basename(
-            os.path.normpath(REPOSITORY_PATH))
-        if TEST_KEYWORD not in base_repo_path:
-            raise InvalidOperation("Be careful. The repository for the tests "
-                                   "is not a test repository. I will not "
-                                   "empty the database and I will not delete "
-                                   "the repository. Repository path: "
-                                   "{}".format(REPOSITORY_PATH))
-
+    def clean_db(self):
         from aiida.backends.djsite.db.models import DbComputer
 
         # I first delete the workflows
@@ -129,6 +97,23 @@ class AiidaTestCase(unittest.TestCase):
         from aiida.backends.djsite.db.models import DbLog
 
         DbLog.objects.all().delete()
+
+    # Note this is has to be a normal method, not a class method
+    def tearDownClass_method(self):
+        from aiida.settings import REPOSITORY_PATH
+        from aiida.common.setup import TEST_KEYWORD
+        from aiida.common.exceptions import InvalidOperation
+
+        base_repo_path = os.path.basename(
+            os.path.normpath(REPOSITORY_PATH))
+        if TEST_KEYWORD not in base_repo_path:
+            raise InvalidOperation("Be careful. The repository for the tests "
+                                   "is not a test repository. I will not "
+                                   "empty the database and I will not delete "
+                                   "the repository. Repository path: "
+                                   "{}".format(REPOSITORY_PATH))
+
+        self.clean_db()
 
         # I clean the test repository
         shutil.rmtree(REPOSITORY_PATH, ignore_errors=True)
